refactor(recording): move video validation policy
